Reorganize units/unit.rs to be an internal module with the parent re-exporting names
